pyproject.toml
[tool.poetry]
name = "aws-autorecycle"
version = "0.0.0"
description = "Trigger autorecycle stepfunction"
authors = ["HMRC Digital"]

[[tool.poetry.source]]
name = "artifactory"
url = "https://artefacts.tax.service.gov.uk/artifactory/api/pypi/pips/simple/"

[tool.poetry.dependencies]
python = "^3.9"
boto3 = "^1.16.38"
aws-lambda-logging = "^0.1.1"
pytz = "^2023.3"
pydantic = "^2.0.2"
aws-lambda-powertools = "^2.26.0"
ansicolors = "^1.1.8"
pymongo = "^4.5.0"
aws-get-vault-object = "^0.0.2"
urllib3 = ">=1.25.4,<1.27"
tenacity = "^8.2.3"

[tool.poetry.dev-dependencies]
pytest = "^7.2.1"
mock = "2.0.0"
coverage = "4.3.4"
parameterized = "0.7.4"
bandit = "^1.6.2"
pre-commit = "^2.1.1"
reorder-python-imports = "^2.0.0"
safety = "^2.3.5"
version-incrementor = "^1.6.0"
freezegun = "^1.2.2"
types-pytz = "^2023.3.0.0"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
